pyproject.toml
[tool.poetry]
name = "practice-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
pymongo = "^3.11.3"
dnspython = "^2.1.0"
discord-pretty-help = "^1.2.2"
praw = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
